# Faster performance for decode function