--- a/.github/workflows/ci.yml
+++ b/.github/workflows/ci.yml
@@ -14,7 +14,7 @@ jobs:
 
     strategy:
       matrix:
-        node-version: [12.x, 14.x]
+        node-version: [12.x, 14.x, 16.x]
 
     steps:
     # Checks-out your repository under $GITHUB_WORKSPACE, so your job can access it
@@ -31,3 +31,6 @@ jobs:
 
     - name: Run tests
       run: npm test
+
+    - name: Run benchmarks
+      run: node benchmark/benchmark.mjs