--- a/README.md
+++ b/README.md
@@ -4,24 +4,23 @@
 
 Topobuf is a compact binary encoding for topographic data.
 
-Topobuf provides _lossless_ compression of TopoJSON data into [protocol buffers](https://developers.google.com/protocol-buffers/).
+Topobuf provides _lossless_[^1] compression of [TopoJSON](https://github.com/topojson/topojson) data into [protocol buffers](https://developers.google.com/protocol-buffers/).
 Advantages over using JSON-based formats alone:
 
 - **Very compact**: typically makes TopoJSON 2-3 times smaller.
 - Smaller even when comparing gzipped sizes: 20-30% for TopoJSON.
+- Can store topology objects too large for `JSON.stringify` / `JSON.parse`
 - Can accommodate any TopoJSON data, including extensions with arbitrary properties.
 
 #### Sample compression sizes
-|                   | normal    | gzipped  |
-|-------------------|-----------|----------|
-| us-zips.json      | 101.85 MB | 26.67 MB |
-| us-zips.pbf       | 12.24 MB  | 10.48 MB |
-| us-zips.topo.json | 15.02 MB  | 3.19 MB  |
-| us-zips.topo.pbf  | 4.85 MB   | 2.72 MB  |
-| idaho.json        | 10.92 MB  | 2.57 MB  |
-| idaho.pbf         | 1.37 MB   | 1.17 MB  |
-| idaho.topo.json   | 1.9 MB    | 612 KB   |
-| idaho.topo.pbf    | 567 KB    | 479 KB   |
+|                       | normal   | gzipped |
+|-----------------------|----------|---------|
+| pa-census-blocks.json | 280 MB   | 44 MB   |
+| pa-census-blocks.pbf  | 114 MB   | 35 MB   |
+| us-zips.json          | 15.02 MB | 3.19 MB |
+| us-zips.pbf           | 4.85 MB  | 2.72 MB |
+| idaho.json            | 1.9 MB   | 612 KB  |
+| idaho.pbf             | 567 KB   | 479 KB  |
 
 
 ## Install
@@ -52,3 +51,5 @@ Given a [Pbf](https://github.com/mapbox/pbf) object with topobuf data, return a 
 ## See more
 
 This library is based on [`geobuf`](https://github.com/mapbox/geobuf) by Mapbox, which provides similar functionality for GeoJSON
+
+[^1]: When using [quantized TopoJSON](https://github.com/topojson/topojson-client/blob/master/README.md#quantize) - _nearly lossless_ otherwise