--- a/package.json
+++ b/package.json
@@ -26,13 +26,14 @@
   },
   "homepage": "https://github.com/azavea/topobuf",
   "devDependencies": {
+    "@thi.ng/bench": "^3.1.3",
     "jshint": "2.13.4",
     "tap": "^15.1.6"
   },
   "dependencies": {
-    "pbf": "^3.2.1",
     "@types/pbf": "^3.0.2",
-    "@types/topojson-specification": "^1.0.2"
+    "@types/topojson-specification": "^1.0.2",
+    "pbf": "^3.2.1"
   },
   "jshintConfig": {
     "trailing": true,