Techniques applied:
 - reuse values & prevP instead of assigning new arrays
 - use new Array() in cases where length is known ahead of time
 - prefer indexing over using .push() when possible
 - move array.length out of for-loops
 
 Benchmarks:
 DE - this is checked into the repo & run as part of GH Actions
 ```
benchmarking: decode de census blocks (topobuf improved)
        warmup... 1316.69ms (5 runs)
        total: 27744.31ms, runs: 100 (@ 1 calls/iter)
        mean: 277.44ms, median: 262.88ms, range: [237.97..370.11]
        q1: 249.41ms, q3: 311.62ms
        sd: 12.35%
 benchmarking: decode de census blocks (topobuf original)
        warmup... 1950.24ms (5 runs)
        total: 35306.07ms, runs: 100 (@ 1 calls/iter)
        mean: 353.06ms, median: 327.43ms, range: [287.10..504.32]
        q1: 314.84ms, q3: 374.80ms
        sd: 15.31%
benchmarking: decode de census blocks (v8)
        warmup... 1342.38ms (5 runs)
        total: 28516.00ms, runs: 100 (@ 1 calls/iter)
        mean: 285.16ms, median: 214.96ms, range: [179.42..500.37]
        q1: 205.45ms, q3: 419.76ms
        sd: 39.83%
benchmarking: JSON.parse
        warmup... 976.45ms (5 runs)
        total: 19386.82ms, runs: 100 (@ 1 calls/iter)
        mean: 193.87ms, median: 184.75ms, range: [160.75..272.58]
        q1: 172.84ms, q3: 218.48ms
        sd: 13.88%
```

PA was too big to check into git, so I just tested it locally

```
benchmarking: decode pa census blocks (topobuf improved)
        warmup... 29058.86ms (3 runs)
        total: 95192.60ms, runs: 15 (@ 1 calls/iter)
        mean: 6346.17ms, median: 6339.94ms, range: [6155.39..6574.00]
        q1: 6294.47ms, q3: 6487.87ms
        sd: 1.96%
benchmarking: decode pa census blocks (topobuf original)
        warmup... 32722.91ms (3 runs)
        total: 163888.58ms, runs: 15 (@ 1 calls/iter)
        mean: 10925.91ms, median: 10959.38ms, range: [10698.45..11060.15]
        q1: 10838.83ms, q3: 11033.05ms
        sd: 1.00%
benchmarking: decode pa census blocks (v8)
        warmup... 25402.78ms (3 runs)
        total: 130589.64ms, runs: 15 (@ 1 calls/iter)
        mean: 8705.98ms, median: 8657.52ms, range: [8445.19..8978.51]
        q1: 8625.25ms, q3: 8879.77ms
        sd: 1.71%
benchmarking: JSON.parse
        warmup... 17742.65ms (3 runs)
        total: 74554.91ms, runs: 15 (@ 1 calls/iter)
        mean: 4970.33ms, median: 5159.99ms, range: [4485.36..5388.14]
        q1: 4686.48ms, q3: 5264.77ms
        sd: 5.95%
```